修复：应用 Element Plus 主题变量
