+++ src/styles/element-variables.scss
@@ -8,18 +8,21 @@ $--color-primary: #1890ff;
 $--color-success: #13ce66;
 $--color-warning: #FFBA00;
 $--color-danger: #ff4949;
-// $--color-info: #1E1E1E;
 
 $--button-font-weight: 400;
-
-// $--color-text-regular: #1f2d3d;
-
 $--border-color-light: #dfe4ed;
 $--border-color-lighter: #e6ebf5;
+$--table-border: 1px solid #dfe6ec;
 
-$--table-border:1px solid#dfe6ec;
-
-/* Element Plus uses CSS variables by default; keep custom color tokens here. */
+/* Element Plus uses CSS variables by default */
+:root {
+  --el-color-primary: #{$--color-primary};
+  --el-color-success: #{$--color-success};
+  --el-color-warning: #{$--color-warning};
+  --el-color-danger: #{$--color-danger};
+  --el-border-color-light: #{$--border-color-light};
+  --el-border-color-lighter: #{$--border-color-lighter};
+}
 
 // the :export directive is the magic sauce for webpack
 // https://www.bluematador.com/blog/how-to-share-variables-between-js-and-sass
